+++ praudio/__init__.py
@@ -1,0 +1,6 @@
+"""Logger gets initialised here."""
+
+import logging
+import sys
+
+logging.basicConfig(stream=sys.stdout, level=logging.INFO)