+++ praudio/errors.py
@@ -1,0 +1,4 @@
+"""This module contains custom errors."""
+
+class FileExtensionError(Exception):
+    """Error that is thrown when the extension of a file isn't allowed."""