+++ test/test_dummy.py
@@ -1,4 +1,0 @@
-from praudio.dummy import sum
-
-def test_sum():
-    assert sum(1, 2) == 3