+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 This file provides a history of the changes in the repo.
 
+# [0.0.4] - 21.08.21 - Valerio Velardo
+- Implemented Loader object
+- Implemented Signal object
+- Implemented FileExtensionError
+- Implemented unittests for all new objects
+
 # [0.0.3] - 21.08.21 - Valerio Velardo
 - Added Makefile with quick commands for installing, linter, testing, typehint
 - Added mypy and pylint config files